refactor: new src dir (#33)
* refactor: new src dir

* refactor: remove old dir
