+++ nodemon.json
@@ -2,5 +2,5 @@
   "watch": ["src/**/*.ts"],
   "ext": "ts,json",
   "ignore": ["src/**/*.spec.ts"],
-  "exec": "ts-node ./bin/www"
+  "exec": "ts-node ./src/server"
 }
