+++ package.json
@@ -4,7 +4,7 @@
   "license": "MIT",
   "private": true,
   "scripts": {
-    "start": "ts-node ./bin/www",
+    "start": "ts-node ./src/server",
     "dev": "nodemon",
     "debug": "nodemon --inspect",
     "prepare": "husky install"
