+++ src/pages/error.jade
@@ -1,4 +1,4 @@
-extends layout
+extends ../components/layouts/layout
 
 block content
   h1= message
